perform ast traversal for all TokenPassBase descendants only once per translation unit

+++ trollint.py
@@ -9,9 +9,12 @@ import passes.base.pass_base
 from progressbar import progressbar
 from diagnostic import from_clang_diagnostic
 import report
-from itertools import groupby
-from utils import get_clang_args, unique, get_clang_analyzer_diagnostics
+from itertools import groupby, imap, chain
+from utils import get_clang_args, unique
+from utils import get_clang_analyzer_diagnostics, get_children_recursively
 from multiprocessing import Pool
+
+PARALLEL = True
 
 
 def discover_pass_classes():
@@ -55,6 +58,17 @@ def lint_one_file(filename, pass_classes, clang_args):
                and not c.location.file.name.startswith('opt')
                and not c.location.file.name.startswith('./opt')]
 
+    def group_cursors_by_kind(cs):
+        result = {}
+        all_cs = chain(*(get_children_recursively(c) for c in cs))
+        for kind, sub_cs in groupby(all_cs, lambda c: c.kind):
+            if kind not in result:
+                result[kind] = []
+            result[kind] += list(sub_cs)
+        return result
+
+    cursors_by_kind = group_cursors_by_kind(local_cursors)
+
     diags = []
 
     # get clang compiler diagnostics
@@ -75,6 +89,9 @@ def lint_one_file(filename, pass_classes, clang_args):
 
         if 'cursors' in pass_class.needs:
             p.cursors = local_cursors
+
+        if 'cursors_by_kind' in pass_class.needs:
+            p.cursors_by_kind = cursors_by_kind
 
         if 'text' in pass_class.needs:
             p.text = blob
@@ -114,17 +131,22 @@ def lint_files(filenames, pass_classes, clang_args):
     def mk_task(filename):
         return (filename, pass_classes, clang_args)
 
-    # TODO: accept -jN and/or read process count from config
-    pool = Pool(processes=4)
     diags = []
     tasks = (mk_task(f) for f in filenames)
-    try:
-        for r in progressbar(pool.imap(worker, tasks),
-                             length=len(filenames)):
+
+    if PARALLEL:
+        # TODO: accept -jN and/or read process count from config
+        pool = Pool(processes=4)
+        try:
+            for r in progressbar(pool.imap(worker, tasks),
+                                 length=len(filenames)):
+                diags += r
+        except KeyboardInterrupt:
+            pool.terminate()
+            return []
+    else:
+        for r in progressbar(imap(worker, tasks), length=len(filenames)):
             diags += r
-    except KeyboardInterrupt:
-        pool.terminate()
-        return []
 
     diags = map(strip_dot_slash, diags)
     diags = filter(interesting_file, diags)
